refactor(crypto): Split merkle_proof.rs and add tracing
Split the 513-line merkle_proof.rs into a directory with smaller files:
- mod.rs (27 lines) - Module declarations and re-exports
- proof_node.rs (74 lines) - ProofNode struct and methods
- proof.rs (130 lines) - MerkleProof struct and verification
- signed_proof.rs (140 lines) - SignedMerkleProof and signing
- extraction.rs (176 lines) - ProofBlockstore trait and extract_proof

All files now under 200-line preferred limit per project guidelines.

Also:
- Convert 24 eprintln! debug statements to structured tracing::debug!
- Add encryption tracing per issue #338:
  - merge_handler.rs: debug on decrypt success, warn on failure
  - block_builder.rs: debug on field encryption
- Apply cargo fmt formatting
