+++ crates/db/src/auto_commit_mutator.rs
@@ -138,9 +138,10 @@ impl<S: Store + 'static> DocMutator for AutoCommitMutator<S> {
                     let enc_config = get_encryption_config();
                     // Get signing config from thread-local (set by FFI exec_request)
                     let sign_config = get_signing_config();
-                    eprintln!(
-                        "[SIGN-DEBUG] auto_commit_mutator::create sign_config.is_some()={}",
-                        sign_config.is_some()
+                    tracing::debug!(
+                        has_signing_config = sign_config.is_some(),
+                        has_encryption_config = enc_config.is_some(),
+                        "Auto-commit create mutation configs"
                     );
 
                     // For create operations, all fields are new - pass None for modified_fields
