+++ crates/db/src/lensed_auto_commit_fetcher.rs
@@ -752,10 +752,9 @@ impl<S: Store + 'static> DocFetcher for LensedAutoCommitFetcher<S> {
                     if has_full_key {
                         let mut key_values = vec![value.clone()];
                         key_values.extend(suffix_values.iter().cloned());
-                        let mut iter =
-                            index.get(&datastore, &key_values).await.map_err(|e| {
-                                query::error::QueryError::execution(format!("index error: {}", e))
-                            })?;
+                        let mut iter = index.get(&datastore, &key_values).await.map_err(|e| {
+                            query::error::QueryError::execution(format!("index error: {}", e))
+                        })?;
                         let entries = iter.collect_all().await.map_err(|e| {
                             query::error::QueryError::execution(format!(
                                 "index iteration error: {}",
