+++ crates/lens/src/pipeline.rs
@@ -383,16 +383,13 @@ async fn apply_transform(
         store.transform(transform_id, input_stream)?
     };
 
-    let result = output_stream
-        .next()
-        .await
-        .ok_or_else(|| {
-            warn!(
-                transform_id = %transform_id,
-                "Transform produced no output"
-            );
-            Error::Pipeline("transform produced no output".to_string())
-        })?;
+    let result = output_stream.next().await.ok_or_else(|| {
+        warn!(
+            transform_id = %transform_id,
+            "Transform produced no output"
+        );
+        Error::Pipeline("transform produced no output".to_string())
+    })?;
 
     match &result {
         Ok(output_doc) => {
